+++ src/main/java/com/web2/movelcontrol/Model/PessoaJuridica.java
@@ -14,7 +14,7 @@ public class PessoaJuridica extends Pessoa {
 
     public PessoaJuridica() {}
 
-    public PessoaJuridica(int id, String nome, String telefone, String email, String endereco, String cnpj) {
+    public PessoaJuridica(Long id, String nome, String telefone, String email, String endereco, String cnpj) {
         super(id, nome, telefone, email, endereco);
         this.cnpj = cnpj;
         this.tipo = "juridica";
